ai panic handling

+++ bindings/dart/lib/src/web/reranker/ai.dart
@@ -14,6 +14,8 @@ import 'package:xayn_ai_ffi_dart/src/common/reranker/ai.dart' as common
 import 'package:xayn_ai_ffi_dart/src/common/reranker/analytics.dart'
     show Analytics;
 import 'package:xayn_ai_ffi_dart/src/common/reranker/mode.dart' show RerankMode;
+import 'package:xayn_ai_ffi_dart/src/common/result/error.dart'
+    show Code, XaynAiException;
 import 'package:xayn_ai_ffi_dart/src/common/result/outcomes.dart'
     show RerankingOutcomes;
 import 'package:xayn_ai_ffi_dart/src/common/utils.dart' show ToJson;
@@ -30,6 +32,7 @@ const int kReceiveTimeoutInSec = 10;
 /// The Xayn AI.
 class XaynAi implements common.XaynAi {
   final Worker? _worker;
+  bool didPanic = false;
 
   /// Creates and initializes the Xayn AI from a given state.
   ///
@@ -63,9 +66,14 @@ class XaynAi implements common.XaynAi {
       serialized,
     );
 
-    final response = await _call(worker, Method.create, params: params);
-    if (response.isException()) {
-      throw response.exception!;
+    try {
+      final response = await _call(worker, Method.create, params: params);
+      if (response.isException()) {
+        throw response.exception!;
+      }
+    } catch (_) {
+      worker.terminate();
+      rethrow;
     }
 
     return XaynAi._(worker);
@@ -97,6 +105,7 @@ class XaynAi implements common.XaynAi {
     );
 
     if (response.isException()) {
+      checkIfAiPanicked(response.exception!);
       throw response.exception!;
     }
 
@@ -113,6 +122,7 @@ class XaynAi implements common.XaynAi {
     final response = await _call(_worker!, Method.serialize);
 
     if (response.isException()) {
+      checkIfAiPanicked(response.exception!);
       throw response.exception!;
     }
 
@@ -131,6 +141,7 @@ class XaynAi implements common.XaynAi {
     final response = await _call(_worker!, Method.faults);
 
     if (response.isException()) {
+      checkIfAiPanicked(response.exception!);
       throw response.exception!;
     }
 
@@ -147,6 +158,7 @@ class XaynAi implements common.XaynAi {
     final response = await _call(_worker!, Method.analytics);
 
     if (response.isException()) {
+      checkIfAiPanicked(response.exception!);
       throw response.exception!;
     }
 
@@ -163,6 +175,7 @@ class XaynAi implements common.XaynAi {
     final response = await _call(_worker!, Method.syncdataBytes);
 
     if (response.isException()) {
+      checkIfAiPanicked(response.exception!);
       throw response.exception!;
     }
 
@@ -183,6 +196,7 @@ class XaynAi implements common.XaynAi {
     );
 
     if (response.isException()) {
+      checkIfAiPanicked(response.exception!);
       throw response.exception!;
     }
   }
@@ -190,11 +204,29 @@ class XaynAi implements common.XaynAi {
   /// Frees the memory.
   @override
   Future<void> free() async {
-    final response = await _call(_worker!, Method.free);
-    _worker!.terminate();
+    if (didPanic) {
+      // When the wasm module panicked we don't call `free` on it because
+      // it will result in a wasm runtime error.
+      _worker!.terminate();
+      return;
+    }
 
-    if (response.isException()) {
-      throw response.exception!;
+    // the clean path of freeing the wasm module and terminating the worker.
+    try {
+      final response = await _call(_worker!, Method.free);
+      if (response.isException()) {
+        throw response.exception!;
+      }
+    } catch (_) {
+      rethrow;
+    } finally {
+      _worker!.terminate();
+    }
+  }
+
+  void checkIfAiPanicked(XaynAiException exception) {
+    if (exception.code == Code.panic) {
+      didPanic = true;
     }
   }
 }
